+++ activestorage/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Rails 6.0.0.alpha (Unreleased) ##
 
+*   Add support for root-relative path URLs to the disk service.
+
+    Fixes #32129.
+
+    *Andrew White*
+
 *   Rails 6 requires Ruby 2.4.1 or newer.
 
     *Jeremy Daer*